+++ compiler/hash-ast/Cargo.toml
@@ -11,7 +11,6 @@ crossbeam-channel = "0.5.1"
 num_cpus = "1.13.0"
 rayon = "1.5.0"
 thiserror = "1.0"
-num = "0.4"
 closure = "0.3.0"
 hash-utils = { path = "../hash-utils" }
 dashmap = "4.0"
