+++ compiler/hash-ast/src/parse.rs
@@ -1,23 +1,22 @@
 //! Hash compiler module for converting from tokens to an AST tree
 //!
 //! All rights reserved 2021 (c) The Hash Language authors
+use crate::module::ModuleIdx;
+use crate::resolve::ModuleResolver;
 use crate::{
     ast::{self, *},
     error::{ParseError, ParseResult},
-    location::SourceLocation,
+    module::Modules,
 };
 use closure::closure;
-use crossbeam_channel::{unbounded, Sender};
-use hash_utils::counter;
-use log::{debug, log_enabled, Level};
+use crossbeam_channel::unbounded;
+use log::debug;
+use std::sync::atomic::AtomicUsize;
 use std::{
-    collections::{HashMap, HashSet},
     fs,
     path::{Path, PathBuf},
     sync::atomic::Ordering,
-    time::Instant,
 };
-use std::{sync::atomic::AtomicUsize, time::Duration};
 
 #[derive(Debug, Copy, Clone)]
 enum EntryPoint<'a> {
@@ -44,7 +43,7 @@ pub struct ParParser<B> {
     backend: B,
 }
 
-enum ParMessage {
+pub enum ParMessage {
     ModuleImport {
         filename: PathBuf,
         parent: Option<ModuleIdx>,
@@ -52,6 +51,9 @@ enum ParMessage {
     },
     ParsedModule {
         node: ast::Module,
+        index: ModuleIdx,
+    },
+    ReadContents {
         filename: PathBuf,
         contents: String,
     },
@@ -95,8 +97,16 @@ where
             let (s, r) = unbounded::<ParMessage>();
 
             senders.fetch_add(1, Ordering::SeqCst);
-            // spawn the initial job
-            let mut resolver = ParModuleResolver::new(s.clone(), None, directory.to_owned());
+
+            // if this is interactive mode we essentially have to state the filename is '<interactive>'..
+            let contents = match entry {
+                EntryPoint::Module { filename } => {
+                    fs::read_to_string(filename).map_err(|e| (e, filename.to_owned()))?
+                }
+                EntryPoint::Interactive { contents } => contents.to_owned(),
+            };
+
+            let mut resolver = ModuleResolver::new(s.clone(), contents, None, directory.to_owned());
 
             let interactive = match entry {
                 EntryPoint::Module { filename } => {
@@ -129,21 +139,36 @@ where
                             let root_dir = filename.parent().unwrap().to_owned();
                             let s = s.clone();
                             senders.fetch_add(1, Ordering::SeqCst);
+
                             scope.spawn(closure!(ref senders, |_| {
-                                let mut resolver = ParModuleResolver::new(s, parent, root_dir);
-                                resolver.parse_file(filename, &self.backend);
-                                senders.fetch_sub(1, Ordering::SeqCst);
+                                // read the file here and then pass a reference to the resolver with it's file name
+                                let contents = fs::read_to_string(filename)
+                                    .map_err(|e| (e, filename.to_owned()));
+
+                                match contents {
+                                    Ok(contents) => {
+                                        s.try_send(ParMessage::ReadContents { filename, contents })
+                                            .unwrap();
+
+                                        // let mut resolver = ModuleResolver::new(s, &source, parent, root_dir);
+
+                                        resolver.parse_file(&filename, &self.backend);
+                                        senders.fetch_sub(1, Ordering::SeqCst);
+                                    }
+                                    Err(e) => {
+                                        s.try_send(ParMessage::Error(ParseError::from(e))).unwrap();
+                                    }
+                                }
                             }));
                         } else {
                             continue;
                         }
                     }
-                    Ok(ParMessage::ParsedModule {
-                        node,
-                        filename,
-                        contents,
-                    }) => {
-                        modules.add_module(filename, node, contents);
+                    Ok(ParMessage::ParsedModule { node, index }) => {
+                        modules.set_node(index, node);
+                    }
+                    Ok(ParMessage::ReadContents { filename, contents }) => {
+                        modules.add_module(filename, contents);
                     }
                     Ok(ParMessage::Error(e)) => {
                         break Err(e);
@@ -152,8 +177,6 @@ where
                         if senders.load(Ordering::SeqCst) == 0 {
                             // All senders disconnected
                             break Ok(interactive);
-                        } else {
-                            continue;
                         }
                     }
                 }
@@ -172,7 +195,7 @@ where
         &self,
         filename: impl AsRef<Path>,
         directory: impl AsRef<Path>,
-    ) -> ParseResult<Modules> {
+    ) -> ParseResult<Module> {
         let filename = filename.as_ref();
         let directory = directory.as_ref();
         let entry = EntryPoint::Module { filename };
@@ -192,405 +215,12 @@ where
     }
 }
 
-#[inline(always)]
-pub fn timed<T>(op: impl FnOnce() -> T, level: log::Level, on_elapsed: impl FnOnce(Duration)) -> T {
-    if log_enabled!(level) {
-        let begin = Instant::now();
-        let result = op();
-        on_elapsed(begin.elapsed());
-        result
-    } else {
-        op()
-    }
-}
-
 pub trait ParserBackend: Sync {
-    fn parse_module(
-        &self,
-        resolver: &mut impl ModuleResolver,
-        path: &Path,
-        contents: &str,
-    ) -> ParseResult<ast::Module>;
+    fn parse_module(&self, resolver: &mut ModuleResolver, path: &Path) -> ParseResult<ast::Module>;
 
     fn parse_interactive(
         &self,
-        resolver: &mut impl ModuleResolver,
+        resolver: &mut ModuleResolver,
         contents: &str,
     ) -> ParseResult<AstNode<ast::BodyBlock>>;
-}
-
-pub trait ModuleResolver {
-    fn add_module(
-        &mut self,
-        import_path: impl AsRef<Path>,
-        location: Option<SourceLocation>,
-    ) -> ParseResult<ModuleIdx>;
-}
-
-fn parse_file(
-    resolved_filename: impl AsRef<Path>,
-    resolver: &mut impl ModuleResolver,
-    backend: &impl ParserBackend,
-) -> Result<(ast::Module, String), ParseError> {
-    debug!("Parsing file: {:?}", resolved_filename.as_ref());
-
-    // load the file in...
-    let source = fs::read_to_string(resolved_filename.as_ref())
-        .map_err(|e| (e, resolved_filename.as_ref().to_owned()))?;
-
-    let module = timed(
-        || backend.parse_module(resolver, resolved_filename.as_ref(), &source),
-        Level::Debug,
-        |elapsed| debug!("ast: {:.2?}", elapsed),
-    )?;
-
-    Ok((module, source))
-}
-
-struct ParModuleResolver {
-    sender: Sender<ParMessage>,
-    parent: Option<ModuleIdx>,
-    root_dir: PathBuf,
-}
-
-impl ParModuleResolver {
-    pub fn new(sender: Sender<ParMessage>, parent: Option<ModuleIdx>, root_dir: PathBuf) -> Self {
-        Self {
-            sender,
-            parent,
-            root_dir,
-        }
-    }
-
-    pub fn add_module_send_error(
-        &mut self,
-        import_path: impl AsRef<Path>,
-        location: Option<SourceLocation>,
-    ) -> Option<ModuleIdx> {
-        match self.add_module(import_path, location) {
-            Ok(index) => Some(index),
-            Err(err) => {
-                self.sender.try_send(ParMessage::Error(err)).unwrap();
-                None
-            }
-        }
-    }
-
-    fn parse_file(&mut self, resolved_filename: impl AsRef<Path>, backend: &impl ParserBackend) {
-        let parse_result = parse_file(&resolved_filename, self, backend);
-
-        let message = match parse_result {
-            Ok((node, source)) => ParMessage::ParsedModule {
-                node,
-                contents: source,
-                filename: resolved_filename.as_ref().to_owned(),
-            },
-            Err(err) => ParMessage::Error(err),
-        };
-
-        self.sender.try_send(message).unwrap();
-    }
-}
-
-impl ModuleResolver for ParModuleResolver {
-    fn add_module(
-        &mut self,
-        import_path: impl AsRef<Path>,
-        location: Option<SourceLocation>,
-    ) -> ParseResult<ModuleIdx> {
-        let resolved_path = resolve_path(import_path, &self.root_dir, location)?;
-        let index = ModuleIdx::new();
-
-        self.sender
-            .try_send(ParMessage::ModuleImport {
-                index,
-                filename: resolved_path,
-                parent: self.parent,
-            })
-            .unwrap();
-
-        Ok(index)
-    }
-}
-
-counter! {
-    name: ModuleIdx,
-    counter_name: MODULE_COUNTER,
-    visibility: pub,
-    method_visibility:,
-}
-
-/// Represents a set of loaded modules.
-#[derive(Debug, Default)]
-pub struct Modules {
-    indexes: HashSet<ModuleIdx>,
-    path_to_index: HashMap<PathBuf, ModuleIdx>,
-    filenames_by_index: HashMap<ModuleIdx, PathBuf>,
-    modules_by_index: HashMap<ModuleIdx, ast::Module>,
-    contents_by_index: HashMap<ModuleIdx, String>,
-    deps_by_index: HashMap<ModuleIdx, HashSet<ModuleIdx>>,
-    entry_point: Option<ModuleIdx>,
-}
-
-impl Modules {
-    /// Create a new [Modules] object
-    pub fn new() -> Self {
-        Modules::default()
-    }
-
-    pub fn has_path(&self, path: impl AsRef<Path>) -> bool {
-        self.path_to_index.contains_key(path.as_ref())
-    }
-
-    pub fn has_entry_point(&self) -> bool {
-        self.entry_point.is_some()
-    }
-
-    pub fn get_entry_point(&self) -> Option<Module<'_>> {
-        Some(self.get_by_index(self.entry_point?))
-    }
-
-    pub fn get_entry_point_unchecked(&self) -> Module<'_> {
-        self.get_entry_point().unwrap()
-    }
-
-    pub fn get_by_index(&self, index: ModuleIdx) -> Module<'_> {
-        Module {
-            index,
-            modules: self,
-        }
-    }
-
-    pub fn get_by_path(&self, path: impl AsRef<Path>) -> Option<Module<'_>> {
-        self.path_to_index
-            .get(path.as_ref())
-            .map(|&idx| self.get_by_index(idx))
-    }
-
-    pub fn get_by_path_unchecked(&self, path: impl AsRef<Path>) -> Module<'_> {
-        self.get_by_path(path).unwrap()
-    }
-
-    fn add_module(&mut self, path: PathBuf, node: ast::Module, contents: String) -> Module<'_> {
-        let index = self.reserve_index();
-        self.add_module_at(index, path, node, contents)
-    }
-
-    fn add_module_at(
-        &mut self,
-        index: ModuleIdx,
-        path: PathBuf,
-        node: ast::Module,
-        contents: String,
-    ) -> Module<'_> {
-        self.path_to_index.insert(path.clone(), index);
-        self.filenames_by_index.insert(index, path);
-        self.modules_by_index.insert(index, node);
-        self.contents_by_index.insert(index, contents);
-
-        Module {
-            index,
-            modules: self,
-        }
-    }
-
-    fn reserve_index(&mut self) -> ModuleIdx {
-        let next = ModuleIdx::new();
-        self.indexes.insert(next);
-        self.deps_by_index.insert(next, HashSet::new());
-        next
-    }
-
-    fn set_entry_point(&mut self, index: ModuleIdx) {
-        self.entry_point = Some(index);
-    }
-
-    fn add_dependency(&mut self, parent: ModuleIdx, child: ModuleIdx) {
-        self.deps_by_index.get_mut(&parent).unwrap().insert(child);
-    }
-
-    pub fn iter(&self) -> impl Iterator<Item = Module<'_>> {
-        self.filenames_by_index.keys().map(move |&index| Module {
-            index,
-            modules: self,
-        })
-    }
-}
-
-/// Represents a single module.
-pub struct Module<'modules> {
-    index: ModuleIdx,
-    modules: &'modules Modules,
-}
-
-impl<'modules> Module<'modules> {
-    pub fn all_modules(&self) -> &Modules {
-        self.modules
-    }
-
-    pub fn index(&self) -> ModuleIdx {
-        self.index
-    }
-
-    pub fn is_entry_point(&self) -> bool {
-        self.modules.entry_point == Some(self.index)
-    }
-
-    pub fn ast(&self) -> &ast::Module {
-        &self.modules.modules_by_index.get(&self.index).unwrap()
-    }
-
-    pub fn content(&self) -> &str {
-        self.modules
-            .contents_by_index
-            .get(&self.index)
-            .unwrap()
-            .as_ref()
-    }
-
-    pub fn dependencies(&'modules self) -> impl Iterator<Item = Module<'modules>> {
-        self.modules
-            .deps_by_index
-            .get(&self.index)
-            .unwrap()
-            .iter()
-            .map(move |&index| Module {
-                index,
-                modules: self.modules,
-            })
-    }
-
-    pub fn filename(&self) -> &Path {
-        self.modules
-            .filenames_by_index
-            .get(&self.index)
-            .unwrap()
-            .as_ref()
-    }
-}
-
-/// The location of a build directory of this package, this used to resolve where the standard library
-/// is located at.
-static BUILD_DIR: &str = env!("CARGO_MANIFEST_DIR");
-
-/// Name of the prelude module
-static PRELUDE: &str = "prelude";
-
-// @@FIXME: this is what we should be looking at rather than doing at runtime!
-// Module names that are used within the standard library
-// const MODULES: &[&Path] = get_stdlib_modules!("./stdlib");
-//
-/// Function that builds a module map of the standard library that is shipped
-/// with the compiler distribution. Standard library modules are referenced
-/// within imports
-pub fn get_stdlib_modules(dir: impl AsRef<Path>) -> Vec<PathBuf> {
-    let mut paths: Vec<PathBuf> = Vec::new();
-
-    if dir.as_ref().is_dir() {
-        for entry in fs::read_dir(dir).unwrap() {
-            match entry {
-                Ok(p) => {
-                    let path = p.path();
-
-                    if path.is_dir() {
-                        // recurse and get all of the files with the prefix
-                        let prefix: &Path = path.file_stem().unwrap().as_ref();
-
-                        for entry in get_stdlib_modules(path.as_path()) {
-                            paths.push(prefix.join(entry));
-                        }
-                    } else if path.is_file() {
-                        let file_name = path.file_stem().unwrap();
-
-                        // Special case, don't add prelude to the module list since we don't want to allow
-                        // it to be imported under the normal standard library imports.
-                        if file_name == PRELUDE {
-                            continue;
-                        }
-
-                        // This shouldn't happen but if there is a file which does not have a hash extension, we shouldn't add it
-                        if path.extension().unwrap_or_default() != "hash" {
-                            continue;
-                        }
-
-                        paths.push(PathBuf::from(file_name));
-                    }
-                }
-                Err(e) => panic!("Unable to read standard library folder: {}", e),
-            }
-        }
-    }
-
-    paths
-}
-
-pub fn resolve_path(
-    path: impl AsRef<Path>,
-    wd: impl AsRef<Path>,
-    location: Option<SourceLocation>,
-) -> ParseResult<PathBuf> {
-    let path = path.as_ref();
-    let wd = wd.as_ref();
-
-    let stdlib_path: PathBuf = [BUILD_DIR, "..", "stdlib"].iter().collect();
-    let modules = get_stdlib_modules(stdlib_path);
-
-    // check if the given path is equal to any of the standard library paths
-    if modules.contains(&path.to_path_buf()) {
-        return Ok(path.to_path_buf());
-    }
-
-    // otherwise, we have to resolve the module path based on the working directory
-    let work_dir = wd.canonicalize().unwrap();
-    let raw_path = work_dir.join(path);
-
-    // If the provided path is a directory, we assume that the user is referencing an index
-    // module that is located within the given directory. This takes precendence over checking
-    // if a module is named that directory.
-    // More info on this topic: https://hash-org.github.io/lang/modules.html#importing
-    if raw_path.is_dir() {
-        let idx_path = raw_path.join("index.hash");
-
-        if idx_path.exists() {
-            return Ok(idx_path);
-        }
-
-        // ok now check if the user is referencing a module instead of the dir
-        let raw_path_hash = raw_path.with_extension("hash");
-        if raw_path_hash.exists() {
-            return Ok(raw_path_hash);
-        }
-
-        // @@Copied
-        Err(ParseError::ImportError {
-            import_name: path.to_path_buf(),
-            src: location.unwrap_or_else(SourceLocation::interactive),
-        })
-    } else {
-        // we don't need to anything if the given raw_path already has a extension '.hash',
-        // since we don't dissalow someone to import a module and reference the module with
-        // the name, like so...
-        //
-        // > let lib = import("lib.hash");
-        // same as...
-        // > let lib = import("lib");
-
-        match raw_path.extension() {
-            Some(k) if k == "hash" => Ok(raw_path),
-            _ => {
-                // add hash extension regardless and check if the path exists...
-                let raw_path_hash = raw_path.with_extension("hash");
-
-                // Only try to check this route if the provided file did not already have an extension
-                if raw_path.extension().is_none() && raw_path_hash.exists() {
-                    Ok(raw_path_hash)
-                } else {
-                    Err(ParseError::ImportError {
-                        import_name: path.to_path_buf(),
-                        src: location.unwrap_or_else(SourceLocation::interactive),
-                    })
-                }
-            }
-        }
-    }
 }
